+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Added `toVector`, `toUnboxedVector`.
+
 # 1.1.0
 
 * Added `toMap`, `toSet`, `toIntMap`, `toIntSet`, `toHashMap`, `toHashSet`.